Define SHADOW sneak behavior based on game
The current SHADOW behavior always allows sneaking since this was brought over from Raven's version of it (while also adding a GHOST flag to emulate its fuller behaviors). This has very odd effects in Doom, so it now only applies this behavior to Raven games specifically. A compat flag, compat_zdoomshadow, was added to bring back the old behavior of allowing it in any game.