Made BobWeapon functions server-side
Currently these are play scoped but called from the renderer, breaking its contract. This also causes play scoped fields to be modified after backing up the player for rollback, making it fundamentally impossible to fix in the future. Instead, call it across all players in P_PlayerThink based on last render type (this is a hack but I don't think there's a better way to solve this). The core bobbing is now linearly interpolated to try and emulate the old behavior as best as possible.