Made the mouse less finnicky in menus
Previously this would "wake up" at the slightest amount of motion, making the mouse a detriment if something caused it to do a minor input in the menu. Now it requires a small amount of motion (based on the screen's height) before it will start registering motion inputs regularly if it has been dormant.