Updated vanilla transparency handling
Previously there were two auto options deciding whether or not there was dehacked or DECORATE present. This has been reworked so that only dehacked Actors and Actors in the core ZScript lump are now considered for it as these are the only safe ones. All other Actors should default to transparency since this is likely what they were designed for. Because of this, this option will now default to auto.